Code cleanup for Day 6
Consolidated code. Got rid of the brute force checks of the whole
lab map and just used the positions that the guard could go to for
a good walk through the lab. Way faster! When I was talking to
Scott about part 2, this was my original thought on how to solve
it, but I went for the brute force at first.