Code cleanup for Day 7
Consolidating the parsing of each line into a function. Moved the
operation work into a function. Then realized the solve logic was
the same, except for the number of operators. So, moved it into
its own function. Probably could just call it from the main, but
I like keeping the solve_part1 and solve_part2 function names.