Solution for Day 20
Took me a bit to get this done. Couldn't work on it coninuously.
Actually originally solved part 1 with not understanding the problem
completely. Then started to work on part 2 and didn't think of it
the right way. Thanks to Scott for helping me think of it the right
way. Then had a few problems as I worked through the right way.
But, finally got it solved.

      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 20   15:26:20  17894      0       >24h  19155      0