Solution for Day 22
Started on the 24th. Took a bit for part 2. I tried a brute force
way, which worked for the example, but took a long time and I didn't
get the right answer. I kept working through it, but missed one
thing, after you get the pattern, you don't keep looking. That was
the problem I was having even with the current solution.

      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 22       >24h  22254      0       >24h  19132      0