Solution for Day 23
Started on the 24th. Part 1 wasn't too hard. Part 2 was a little more
difficult. I realized that it can only be related to one computer and
the set of connections from it. The difficult part was figuring out
how to do the intersection of connections.

      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 23       >24h  20409      0       >24h  18214      0