Solution for Day 21
Probably the hardest one overall. I did several different things.
The one I expected to work was to manually enter all the paths from
one key to another. That was error prone. So, I switched to a
programmatic way of getting the paths and caching the results. That
worked, but was slow. Then I started watching HyperNeutrino on YouTube.
I got the current optimization for part 1 from her video and all of
part 2. I had some idea that was a solution, but didn't know how to
implement it.

      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 21       >24h  18231      0       >24h  14877      0